Only wrap selections that are comment lines
I think this is to be expected given the name of the extension and the
documentation of `parse`. I also don’t think that source lines should
necessarily be wrapped at the page guide as opposed to at syntactic breaks
in the line like between arguments in a function call. Finally, this change lays
the foundation for some upcoming fixes.