Describe how to develop and install the extension
We check in the schemes so that these instructions will work. In particular,
we’ve made the following changes to the default schemes:

- Changing XCommentWrap’s Run build configuration to Release appears
necessary to get the extension to show up in System Preferences. I think that
this should be fine because there’s no reason to debug the app itself.
- Changing XCommentWrap-extension’s debug executable to Xcode (instead
of “Ask on launch”) appears necessary to get the debugger to attach,
otherwise some error about a dylib not loading appears. This is also more
convenient.